Moved widgets into widgets folder, HighlightableButton, PR Comments
