+++ src/main/kotlin/org/wycliffeassociates/otter/jvm/app/widgets/card/DefaultStyles.kt
@@ -35,7 +35,6 @@ class DefaultStyles : Stylesheet() {
         val defaultBody by cssclass()
         val defaultMajorLabel by cssclass()
         val defaultMinorLabel by cssclass()
-        val regg by cssclass()
     }
 
     init {
